Eliminate 2013 references in *.2015.vcxproj files
Symptom:

Attempting to compile this solution with VS2015 fails because the
VS2013 toolset is not installed.

Discussion:

- Three of the projects have direct references to the Visual Studio
2013 toolset (one with WinXP extensions).  These were easy to update
directly with the VS2015 IDE. After that, all indication on the various
project property dialog windows indicates that the projects were
correctly configured for VS2015.

- However, the solution still would not compile.

- Seven of projects have various 'Reference Projects', which are
additional, internal, linkages to other inter-dependent projects.
This is in addition to the normal linkage to the projects that are
explicitly part of the solution.

- Within the text of the *.2015.vcxproj files, these 'Reference
Projects' still refer to the older *.2013.vcxproj files, thus requiring
the co-existence of the VS2013 compiler.

Solution

- Three projects were directly updated to use the VS2015 toolset.

- I could find no way to update the other seven projects' references
directly via the VS2015 IDE.  Possibly deleting and re-adding them
would work.  However, by using NotePad++ on each of the xxx.2015.vcxproj
files, I found all literals .2013. and changed them to .2015 - these
occurred in the seven .vcxproj files in this commit.

- The entire solution now compiles correctly with the VS2015 compiler,
without the existence of the VS2013 toolset.